backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
Pillow==10.1.0
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
elasticsearch==8.11.0
django-elasticsearch-dsl==7.3.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
djangorestframework-simplejwt==5.3.0
django-filter==23.4
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
flake8==6.1.0
black==23.11.0
isort==5.12.0